answer

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -9,7 +9,7 @@
 
 int main(void)
 {
-	int i, n = 1, pn = 0, cn;
+	long int i, n = 1, pn = 0, cn;
 
 	for (i = 0; i < 98; i++)
 	{
